pyproject.toml
[tool.poetry]
name = "binprov"
version = "0.1.0"
description = ""
authors = ["William Panlener"]

[tool.poetry.dependencies]
python = "^3.10"
pyelftools = "^0.31"
iced-x86 = "^1.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
